+++ CHANGELOG.md
@@ -11,7 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Support change notification via SNS topic (#267).
 - Support change notification via PubSub topic (#294).
-- Add OnEvent on Azure providers (#297).
+- Support change notification via Service Bus topic (#302).
 - Add provider for AWS Parameter Store (#298).
 
 ## [1.0.0] - 2024-03-16